toquetie cosas para validar que no existan productos ni usuarios en ventas, porque sino no se deberia poder eliminar

+++ src/main/java/com/ob/ob/services/Cliente/ClienteServiceImpl.java
@@ -4,7 +4,10 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import com.ob.ob.entity.Cliente;
 import com.ob.ob.entity.Regular;
@@ -16,7 +19,7 @@ import com.ob.ob.repository.VIPRepository;
 import jakarta.transaction.Transactional;
 
 @Service
-public class ClienteServiceImpl implements ClienteService{
+public class ClienteServiceImpl implements ClienteService {
     @Autowired
     private ClienteRepository cliRepo;
     @Autowired
@@ -24,76 +27,92 @@ public class ClienteServiceImpl implements ClienteService{
     @Autowired
     private RegularRepository regularRepo;
 
-    //region REGULAR
+    // region REGULAR
     @Override
     @Transactional
-    public List<Regular> getAllRegular(){
+    public List<Regular> getAllRegular() {
         return regularRepo.findAll();
     }
 
     @Override
     @Transactional
-    public Optional<Regular> getAregular(int id){
+    public Optional<Regular> getAregular(int id) {
         return regularRepo.findById(id);
     }
 
     @Override
     @Transactional
-    public Regular save(Regular cli){
+    public Regular save(Regular cli) {
         return regularRepo.save(cli);
     }
 
     @Override
     @Transactional
-    public Regular update(Regular cli){
+    public Regular update(Regular cli) {
         return regularRepo.save(cli);
     }
-    //endregion
+    // endregion
 
-    //region VIP
+    // region VIP
     @Override
     @Transactional
-    public List<VIP> getAllVIP(){
+    public List<VIP> getAllVIP() {
         return vipRepo.findAll();
     }
 
     @Override
     @Transactional
-    public Optional<VIP> getAvip(int id){
+    public Optional<VIP> getAvip(int id) {
         return vipRepo.findById(id);
     }
 
     @Override
     @Transactional
-    public VIP saveVIP(VIP cli){
+    public VIP saveVIP(VIP cli) {
         return vipRepo.save(cli);
     }
 
     @Override
     @Transactional
-    public VIP updateVIP(VIP cli){
+    public VIP updateVIP(VIP cli) {
         return vipRepo.save(cli);
     }
-    //endregion
+    // endregion
 
-    //region GLOBAL
-    @Override
-    @Transactional
-    public String delete(int id){
-        cliRepo.deleteById(id);
-        return "Se ha eliminado correctamente";
+    // region GLOBAL
+    // @Override
+    // @Transactional
+    // public String delete(int id){
+    // cliRepo.deleteById(id);
+    // return "Se ha eliminado correctamente";
+    // }
+
+    public boolean existUserenVenta(@PathVariable int userId) {
+        return cliRepo.existsByClienteId(userId);
     }
 
     @Override
     @Transactional
-    public List<Cliente> getAll(){
+    public ResponseEntity<String> delete(int id) {
+        if (existUserenVenta(id)) {
+            return new ResponseEntity<>("Existen datos en ventas de este producto, no es posible eliminar",
+                    HttpStatus.METHOD_NOT_ALLOWED);
+        } else {
+            cliRepo.deleteById(id);
+            return new ResponseEntity<>("Se ha eliminado correctamente", HttpStatus.OK);
+        }
+    }
+
+    @Override
+    @Transactional
+    public List<Cliente> getAll() {
         return cliRepo.findAll();
     }
 
     @Override
     @Transactional
-    public Optional<Cliente> getAclient(int id){
+    public Optional<Cliente> getAclient(int id) {
         return cliRepo.findById(id);
     }
-    //endregion
+    // endregion
 }
